Changes for the first version of turning an LED on and off.
We have a GPIO pin manipulation library in Kotlin composed of a lower level primitives in Gpio.kt and a higher level interface in GpioProcessor.kt.

We have modified the layout and the code of SecondFragment.xml and SecondFragment.kt to present two buttons, "LED On" and "LED Off", which toggle the
output of pin 23 of the 40 pin low power connector from high to low in order to light and unlight an LED on a breadboard.

We are still working on the GPS facility. A forum post found mentioned that the GPS may need an external antenna and be out of doors to properly find a
GPS satellite.