New version working on LED blinking.
We have introduced a new file with a new class that implements a thread for blinking an LED. For this initial test, we have hard coded the rhythm and
the LED will blink only a few times with that rhythm of on and off.

We have modified the layout to better align buttons with the text area for messages in the Second Fragment. And we have added a new button to start
the LED blink thread along with a button listener to start the thread when the button is pressed.

We are still working on the GPS facility. A forum post found mentioned that the GPS may need an external antenna and be out of doors to properly find a
GPS satellite.