rewrote the code in Gpio.kt to provide a base class, GpioFile, which is used to provide a set of sysfs interface primitives to read and write a pseudo file. this GpioFile class is then used by class Gpio, GpioLed, and GpioPwm to implement the necessary interface between the user of the GPIO library and the pseudo files that represent the hardware, the GPIO pins and the user LEDs respectively.
we have also updated the Second Fragment user interface to provide buttons to turn on and off user LED 1 and user LED 2, two of the
four user LEDs located between the two USB connectors on the DragonBoard 410C that are available under Android.

we are still working on class GpioPwm and have not done any testing with this class yet. we are still attempting to discover how the
PWM works under Android. we have found some information about using pin 28 with sysfs for PWM under Linux.

we have done testing of both class Gpio and GpioLed under Android and it appears to be working satisfactorily.