First version of the Native C++ function with JNI to interface with the operating system to perform the poll(2) action needed for an input GPIO using edge triggering. also adding the CMakeLists.txt file needed for properly building the Native C++ library for use with Kotlin.
changes include:
 - new C++ source with a JNI interface with an extern "C" modifier
 - init added to MainActivity.kt with System.loadLibrary() to load Native C++ function
 - new method pinPoll() added to class Gpio to use Native C++ function for polling
 - test harness in SecondFragment.kt and layout

we have modified SecondFragment.kt and its layout so as to add a new button to perform a poll(2) of an input GPIO pin reporting the action and its result by appending text lines to the text window.

we have tested this function on an actual DragonBoard 410C and the poll(2) is return immediately with a POLLERR indicated along with POLLPRI indicated. However the value we retrieve is correct. we are looking into what is going on.

we are pushing the source code up though the poll(2) functionality is not complete to provide a baseline for further development.