small change to PollFileService.cpp to allow testing with an emulator without GPIO pins by using a sleep() rather than using the poll() function on a GPIO pseudo file.
changes to Gpio.kt and the pin polling functionality so that it can be used with coroutines.
we have introduced an observable variable which is toggled between 0 and 1 whenever a poll() is
done and the function returns. This allows the user to start a poll in a coroutine and then use
a listener to know when the poll completes either because the GPIO pin was triggered or the poll()
time out expired.

changes to SecondFragment.kt to test the new functionality with coroutine and observable.